# docs-site/src/main/resources/static/api/v2/openapi.yaml
openapi: 3.0.0
info:
  title: Global Payments API
  description: |
    A comprehensive API for processing various types of payments, managing customers, and handling billing.
    
    ## Authentication
    All API requests require authentication using API keys passed in the `Authorization` header:
    ```
    ```
    
    ## Rate Limiting
    - Standard tier: 1000 requests per minute
    - Premium tier: 5000 requests per minute
    
    ## Idempotency
    POST requests support idempotency keys to prevent duplicate operations.
  version: 2.0.0
  contact:
    name: Global Payments API Support
    email: api-support@globalpayments.com
    url: https://docs.globalpayments.com
servers:
  - url: https://api.globalpayments.com/v2
    description: Production server
  - url: https://sandbox-api.globalpayments.com/v2
    description: Sandbox server
security:
  - bearerAuth: []
paths:
  /payments:
    post:
      summary: Create a Payment
      description: |
        Creates a new payment transaction. Supports multiple payment methods including credit cards, 
        digital wallets, net banking, and buy-now-pay-later options.
        
        **Important Notes:**
        - All amounts are in the smallest currency unit (cents for USD, paise for INR)
        - Use idempotency keys to prevent duplicate charges
        - 3D Secure authentication is automatically triggered for eligible transactions
      operationId: create-payment
      tags: [Payments]
      parameters:
        - name: Idempotency-Key
          in: header
          description: Unique key to prevent duplicate requests
          required: false
          schema:
            type: string
            maxLength: 255
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentRequest'
            examples:
              credit_card_payment:
                summary: Credit Card Payment
                value:
                  amount: 2500
                  currency: "USD"
                  description: "Order #12345 - Premium subscription"
                  merchant_info:
                    merchant_id: "merchant_123"
                    terminal_id: "terminal_456"
                    merchant_category_code: "5734"
                  payment_method:
                    type: "credit_card"
                    card:
                      card_number: "4111111111111111"
                      expiry_month: 12
                      expiry_year: 2025
                      cvc: "123"
                      cardholder_name: "John Doe"
                  customer:
                    customer_id: "cust_123"
                    email: "john@example.com"
                    phone: "+1234567890"
                  capture: true
                  metadata:
                    order_id: "12345"
                    customer_segment: "premium"
      responses:
        '201':
          description: Payment created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentResponse'
              examples:
                successful_payment:
                  summary: Successful Payment
                  value:
                    id: "pay_1234567890"
                    status: "succeeded"
                    amount: 2500
                    currency: "USD"
                    description: "Order #12345 - Premium subscription"
                    created_at: "2024-01-15T10:30:00Z"
                    payment_method:
                      type: "credit_card"
                      card:
                        last4: "1111"
                        brand: "visa"
                        exp_month: 12
                        exp_year: 2025
                    receipt_url: "https://receipts.globalpayments.com/pay_1234567890"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '402':
          $ref: '#/components/responses/PaymentRequired'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
    get:
      summary: List Payments
      description: |
        Retrieves a list of payments with optional filtering and pagination.
        Results are returned in reverse chronological order (newest first).
      operationId: list-payments
      tags: [Payments]
      parameters:
        - name: limit
          in: query
          description: Number of payments to return (1-100)
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
        - name: offset
          in: query
          description: Number of payments to skip
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: status
          in: query
          description: Filter by payment status
          required: false
          schema:
            type: string
            enum: [succeeded, pending, failed, requires_action]
        - name: created_after
          in: query
          description: Filter payments created after this timestamp
          required: false
          schema:
            type: string
            format: date-time
        - name: created_before
          in: query
          description: Filter payments created before this timestamp
          required: false
          schema:
            type: string
            format: date-time
        - name: customer_id
          in: query
          description: Filter by customer ID
          required: false
          schema:
            type: string
      responses:
        '200':
          description: A paginated list of payments
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/PaymentResponse'
                  pagination:
                    $ref: '#/components/schemas/PaginationInfo'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    PaymentRequest:
      type: object
      required: [amount, currency, payment_method]
      properties:
        amount:
          type: integer
          description: Amount in the smallest currency unit (e.g., cents for USD)
          minimum: 1
          maximum: 99999999
        currency:
          type: string
          description: 3-letter ISO currency code
          pattern: '^[A-Z]{3}$'
        description:
          type: string
          description: Description of the payment
          maxLength: 500
        capture:
          type: boolean
          description: Whether to capture the payment immediately
          default: true
        merchant_info:
          $ref: '#/components/schemas/MerchantInfo'
        payment_method:
          $ref: '#/components/schemas/PaymentMethod'
        customer:
          $ref: '#/components/schemas/CustomerDetails'
        shipping_address:
          $ref: '#/components/schemas/Address'
        metadata:
          type: object
          description: Additional metadata for the payment
          additionalProperties: true
    PaymentResponse:
      type: object
      properties:
        id:
          type: string
          description: Unique payment identifier
          pattern: '^pay_[a-zA-Z0-9]+$'
        status:
          type: string
          enum: [succeeded, pending, failed, requires_action, requires_capture, cancelled]
          description: Current status of the payment
        amount:
          type: integer
          description: Payment amount in smallest currency unit
        currency:
          type: string
          description: 3-letter ISO currency code
        description:
          type: string
          description: Payment description
        created_at:
          type: string
          format: date-time
          description: Payment creation timestamp
        updated_at:
          type: string
          format: date-time
          description: Payment last update timestamp
        payment_method:
          $ref: '#/components/schemas/PaymentMethodResponse'
        customer:
          $ref: '#/components/schemas/CustomerDetails'
        receipt_url:
          type: string
          format: uri
          description: URL to the payment receipt
        metadata:
          type: object
          description: Additional metadata
          additionalProperties: true
...
    MerchantInfo:
      type: object
      required: [merchant_id]
      properties:
        merchant_id:
          type: string
          description: Merchant identifier
          maxLength: 50
        terminal_id:
          type: string
          description: Terminal identifier
          maxLength: 50
        merchant_category_code:
          type: string
          description: 4-digit MCC code
          pattern: '^[0-9]{4}$'
        merchant_name:
          type: string
          description: Merchant name
          maxLength: 100
        merchant_url:
          type: string
          format: uri
          description: Merchant website URL
    PaymentMethod:
      type: object
      required: [type]
      properties:
        type:
          type: string
          enum: [credit_card, wallet, netbanking, bnpl, bank_transfer]
          description: Payment method type
        card:
          $ref: '#/components/schemas/CardDetails'
        wallet:
          $ref: '#/components/schemas/WalletInfo'
        bank_transfer:
          $ref: '#/components/schemas/BankTransferInfo'
    PaymentMethodResponse:
      type: object
      properties:
        type:
          type: string
          enum: [credit_card, wallet, netbanking, bnpl, bank_transfer]
          description: Payment method type
        card:
          $ref: '#/components/schemas/CardDetailsResponse'
        wallet:
          $ref: '#/components/schemas/WalletInfo'
    CardDetails:
      type: object
      required: [card_number, expiry_month, expiry_year]
      properties:
        card_number:
          type: string
          description: Card number (PAN)
          pattern: '^[0-9]{13,19}$'
        expiry_month:
          type: integer
          description: Expiration month (1-12)
          minimum: 1
          maximum: 12
        expiry_year:
          type: integer
          description: Expiration year (4-digit)
          minimum: 2024
        cvc:
          type: string
          description: Card verification code
          pattern: '^[0-9]{3,4}$'
        cardholder_name:
          type: string
          description: Name on the card
          maxLength: 100
    CardDetailsResponse:
      type: object
      properties:
        last4:
          type: string
          description: Last 4 digits of the card
        brand:
          type: string
          description: Card brand
          enum: [visa, mastercard, amex, discover, jcb, diners, unionpay]
        exp_month:
          type: integer
          description: Expiration month
        exp_year:
          type: integer
          description: Expiration year
        cardholder_name:
          type: string
          description: Name on the card
        country:
          type: string
          description: Card issuing country
        funding:
          type: string
          enum: [credit, debit, prepaid]
          description: Card funding type
    WalletInfo:
      type: object
      required: [provider]
      properties:
        provider:
          type: string
          enum: [apple_pay, google_pay, paypal, samsung_pay]
          description: Wallet provider
        wallet_token:
          type: string
          description: Wallet token
        device_id:
          type: string
          description: Device identifier
    BankTransferInfo:
      type: object
      required: [account_number, routing_number]
      properties:
        account_number:
          type: string
          description: Bank account number
        routing_number:
          type: string
          description: Bank routing number
        account_type:
          type: string
          enum: [checking, savings]
          description: Account type
        bank_name:
          type: string
          description: Bank name
    CustomerDetails:
      type: object
      properties:
        customer_id:
          type: string
          description: Customer identifier
        email:
          type: string
          format: email
          description: Customer email
        phone:
          type: string
          description: Customer phone
        name:
          type: string
          description: Customer name
        billing_address:
          $ref: '#/components/schemas/Address'
    Address:
      type: object
      properties:
        line1:
          type: string
          description: Address line 1
          maxLength: 200
        line2:
          type: string
          description: Address line 2
          maxLength: 200
        city:
          type: string
          description: City
          maxLength: 100
        state:
          type: string
          description: State/Province
          maxLength: 100
        postal_code:
          type: string
          description: Postal/ZIP code
          maxLength: 20
        country:
          type: string
          description: 2-letter ISO country code
          pattern: '^[A-Z]{2}$'
    PaginationInfo:
      type: object
      properties:
        limit:
          type: integer
          description: Number of items per page
        offset:
          type: integer
          description: Number of items skipped
        total:
          type: integer
          description: Total number of items
        has_more:
          type: boolean
          description: Whether there are more items
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              description: Error code
            message:
              type: string
              description: Error message
            details:
              type: array
              description: Additional error details
              items:
                type: object
                properties:
                  field:
                    type: string
                  message:
                    type: string
        request_id:
          type: string
          description: Unique request identifier
  responses:
    BadRequest:
      description: Bad request - invalid parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            invalid_amount:
              summary: Invalid Amount
              value:
                error:
                  code: "invalid_amount"
                  message: "Amount must be a positive integer"
                  details:
                    - field: "amount"
                      message: "Must be greater than 0"
                request_id: "req_1234567890"
    Unauthorized:
      description: Unauthorized - invalid API key
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            invalid_key:
              summary: Invalid API Key
              value:
                error:
                  code: "invalid_api_key"
                  message: "Invalid API key provided"
                request_id: "req_1234567890"
    PaymentRequired:
      description: Payment required - insufficient funds or payment declined
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            card_declined:
              summary: Card Declined
              value:
                error:
                  code: "card_declined"
                  message: "Your card was declined"
                  details:
                    - field: "payment_method"
                      message: "Card declined by issuer"
                request_id: "req_1234567890"
...
    TooManyRequests:
      description: Too many requests - rate limit exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            rate_limit:
              summary: Rate Limit Exceeded
              value:
                error:
                  code: "rate_limit_exceeded"
                  message: "Too many requests. Please try again later."
                request_id: "req_1234567890"
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            server_error:
              summary: Server Error
              value:
                error:
                  code: "internal_error"
                  message: "An unexpected error occurred"
                request_id: "req_1234567890"